+++ db4e/Panes/MonerodRemote.py
@@ -22,63 +22,68 @@ from db4e.Constants.Labels import (
     RPC_BIND_PORT_LABEL, UPDATE_LABEL, ZMQ_PUB_PORT_LABEL
 )
 
-STATIC_CONTENT = f"This screen allows you to view and edit the deployment "
-STATIC_CONTENT += f"settings for the {MONEROD_REMOTE_LABEL} deployment."
-
 class MonerodRemote(Container):
 
-    def set_data(self, depl_config):
+    instance_input = Input(id="instance_input", restrict=f"[a-zA-Z0-9_\-]*", compact=True)
+    ip_addr_input = Input(id="ip_addr_input", restrict=f"[a-z0-9._\-]*", compact=True)
+    rpc_bind_port_input = Input(id="rpc_bind_port_input", restrict=f"[0-9]*", compact=True)
+    zmq_pub_port_input = Input(id="zmq_pub_port_input", restrict=f"[0-9]*", compact=True)
+    orig_instance_input = Input(id="orig_instance_input", classes="hidden")
 
-        instance = depl_config[INSTANCE_FIELD]
-        ip_addr = depl_config[IP_ADDR_FIELD]
-        rpc_bind_port = depl_config[RPC_BIND_PORT_FIELD]
-        zmq_pub_port = depl_config[ZMQ_PUB_PORT_FIELD]
+    def compose(self):
+        # Remote Monero daemon deployment form
+        STATIC_CONTENT = f"This screen allows you to view and edit the deployment "
+        STATIC_CONTENT += f"settings for the {MONEROD_REMOTE_LABEL} deployment."
 
-        md = Vertical(
+        yield Vertical(
             MarkdownViewer(STATIC_CONTENT, show_table_of_contents=False, classes="form_intro"),
 
             Vertical(
                 Horizontal(
-                    Label(INSTANCE_LABEL, id="monerod_remote_instance_label"),
-                    Input(id="monerod_remote_instance",
-                          restrict=f"[a-zA-Z0-9_\-]*", value=instance, compact=True)),
+                    Label(INSTANCE_LABEL, id="instance_label"),
+                    self.instance_input),
                 Horizontal(
-                    Label(IP_ADDR_LABEL, id="monerod_remote_ip_addr_label"),
-                    Input(id="monerod_remote_ip_addr",
-                          restrict=f"[a-z0-9._\-]*", value=ip_addr, compact=True)),
+                    Label(IP_ADDR_LABEL, id="ip_addr_label"),
+                    self.ip_addr_input),
                 Horizontal(
-                    Label(RPC_BIND_PORT_LABEL, id="monerod_remote_rpc_bind_port_label"),
-                    Input(id="monerod_remote_rpc_bind_port",
-                          restrict=f"[0-9]*", value=str(rpc_bind_port), compact=True)),
+                    Label(RPC_BIND_PORT_LABEL, id="rpc_bind_port_label"),
+                    self.rpc_bind_port_input),
                 Horizontal(
-                    Label(ZMQ_PUB_PORT_LABEL, id="monerod_remote_zmq_pub_port_label"),
-                    Input(id="monerod_remote_zmq_pub_port",
-                          restrict=f"[0-9]*", value=str(zmq_pub_port), compact=True)),
-                id="monerod_remote_update_form"),
+                    Label(ZMQ_PUB_PORT_LABEL, id="zmq_pub_port_label"),
+                    self.zmq_pub_port_input),
+                id="monerod_remote_form"),
 
-            Input(id="monerod_remote_orig_instance", value=instance, classes="hidden"),
+            self.orig_instance_input,
 
             Horizontal(
-                Button(label=UPDATE_LABEL, id="monerod_remote_update_button"),
-                Button(label=DELETE_LABEL, id="monerod_remote_delete_button")
-            ))
-        self.remove_children()
-        self.mount(md)
+                Button(label=UPDATE_LABEL, id="update_button"),
+                Button(label=DELETE_LABEL, id="delete_button"),
+                id="buttons"),
+
+        id="pane")
+
+    def set_data(self, rec):
+
+        self.instance_input.value = rec[INSTANCE_FIELD]
+        self.orig_instance_input.value = rec[INSTANCE_FIELD]
+        self.ip_addr_input.value = rec[IP_ADDR_FIELD]
+        self.rpc_bind_port_input.value = str(rec[RPC_BIND_PORT_FIELD])
+        self.zmq_pub_port_input.value = str(rec[ZMQ_PUB_PORT_FIELD])
         
     def on_button_pressed(self, event: Button.Pressed) -> None:
         event.stop()
         button_id = event.button.id
-        if button_id == "monerod_remote_update_button":
+        if button_id == "update_button":
             form_data = {
                 COMPONENT_FIELD: MONEROD_FIELD,
                 TO_MODULE_FIELD: DEPLOYMENT_MGR_FIELD,
                 TO_METHOD_FIELD: UPDATE_DEPLOYMENT_FIELD,
                 FORM_DATA_FIELD: True,
-                INSTANCE_FIELD: self.query_one("#monerod_remote_instance", Input).value,
-                ORIG_INSTANCE_FIELD: self.query_one("#monerod_remote_orig_instance", Input).value,
-                IP_ADDR_FIELD: self.query_one("#monerod_remote_ip_addr", Input).value,
-                RPC_BIND_PORT_FIELD: self.query_one("#monerod_remote_rpc_bind_port", Input).value,
-                ZMQ_PUB_PORT_FIELD: self.query_one("#monerod_remote_zmq_pub_port", Input).value,
+                INSTANCE_FIELD: self.query_one("#instance_input", Input).value,
+                ORIG_INSTANCE_FIELD: self.query_one("#orig_instance_input", Input).value,
+                IP_ADDR_FIELD: self.query_one("#ip_addr_input", Input).value,
+                RPC_BIND_PORT_FIELD: self.query_one("#rpc_bind_port_input", Input).value,
+                ZMQ_PUB_PORT_FIELD: self.query_one("#zmq_pub_port_input", Input).value,
             }
         else:
             form_data = {
@@ -86,6 +91,6 @@ class MonerodRemote(Container):
                 TO_MODULE_FIELD: DEPLOYMENT_MGR_FIELD,
                 TO_METHOD_FIELD: DELETE_DEPLOYMENT_FIELD,
                 COMPONENT_FIELD: MONEROD_FIELD,
-                INSTANCE_FIELD: self.query_one("#monerod_remote_instance", Input).value,
+                INSTANCE_FIELD: self.query_one("#orig_instance_input", Input).value,
             }
         self.app.post_message(SubmitFormData(self, form_data=form_data))